+++ bcs-app/backend/container_service/infras/host_service/constants.py
