+++ bcs-app/backend/tests/components/test_base.py
@@ -17,7 +17,7 @@ import requests
 from backend.components.base import (
     BaseHttpClient,
     CompInternalError,
-    CompParseResponseError,
+    CompParseBkCommonResponseError,
     CompRequestError,
     response_handler,
     update_request_body,
@@ -119,7 +119,7 @@ class TestBkCommonResponseHandler:
         assert data == expected_data
 
     def test_response_error_data_hander(self):
-        with pytest.raises(CompParseResponseError):
+        with pytest.raises(CompParseBkCommonResponseError):
             response_handler()(self.func_error)()
 
     @pytest.mark.parametrize(
