+++ bcs-app/backend/tests/container_service/infras/host_service/test_host.py
@@ -3,8 +3,8 @@ from unittest.mock import patch
 
 import pytest
 
-from backend.infras.host_service import host
-from backend.infras.host_service.perms import can_use_hosts
+from backend.container_service.infras.host_service import host
+from backend.container_service.infras.host_service.perms import can_use_hosts
 from backend.tests.bcs_mocks.fake_sops import FakeSopsMod, sops_json
 
 fake_cc_host_ok_results = {
@@ -17,32 +17,34 @@ expect_used_ip_list = ["127.0.0.1", "127.0.0.2"]
 
 
 class TestCheckUseHost:
-    @patch("backend.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_ok_results)
+    @patch("backend.container_service.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_ok_results)
     def test_ok(self, biz_id, username):
         assert can_use_hosts(biz_id, username, expect_used_ip_list)
 
-    @patch("backend.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_null_results)
+    @patch("backend.container_service.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_null_results)
     def test_null_resp_failed(self, biz_id, username):
         assert not can_use_hosts(biz_id, username, expect_used_ip_list)
 
-    @patch("backend.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_not_match_results)
+    @patch(
+        "backend.container_service.infras.host_service.perms.get_cc_hosts", return_value=fake_cc_host_not_match_results
+    )
     def test_not_match_failed(self, biz_id, username):
         assert not can_use_hosts(biz_id, username, expect_used_ip_list)
 
 
 class TestGetCCHosts:
-    @patch("backend.infras.host_service.host.cc.get_app_hosts", return_value={"result": False})
+    @patch("backend.container_service.infras.host_service.host.cc.get_app_hosts", return_value={"result": False})
     def test_get_null_hosts(self, mock_get_app_hosts, biz_id, username):
         assert host.get_cc_hosts(username, biz_id) == []
 
-    @patch("backend.infras.host_service.host.cc.get_app_hosts", return_value=fake_cc_host_ok_results)
+    @patch("backend.container_service.infras.host_service.host.cc.get_app_hosts", return_value=fake_cc_host_ok_results)
     def test_get_hosts(self, mock_get_app_hosts, biz_id, username):
         assert host.get_cc_hosts(username, biz_id) == fake_cc_host_ok_results["data"]
 
 
 class TestGetAgentStatus:
     @patch(
-        "backend.infras.host_service.host.gse.get_agent_status",
+        "backend.container_service.infras.host_service.host.gse.get_agent_status",
         return_value=[
             {"ip": "127.0.0.1", "bk_cloud_id": 0, "bk_agent_alive": 1},
             {"ip": "127.0.0.2", "bk_cloud_id": 0, "bk_agent_alive": 1},
@@ -78,7 +80,7 @@ class TestApplyHostApi:
     }
 
     @patch(
-        "backend.infras.host_service.host.sops.SopsClient",
+        "backend.container_service.infras.host_service.host.sops.SopsClient",
         new=FakeSopsMod,
     )
     def test_create_and_start_host_application(self):
@@ -86,7 +88,7 @@ class TestApplyHostApi:
         assert task_id == sops_json.fake_task_id
         assert task_url == sops_json.fake_task_url
 
-    @patch("backend.infras.host_service.host.sops.SopsClient", new=FakeSopsMod)
+    @patch("backend.container_service.infras.host_service.host.sops.SopsClient", new=FakeSopsMod)
     def test_get_task_state_and_steps(self):
         status_and_steps = host.get_task_state_and_steps(sops_json.fake_task_id)
         assert status_and_steps["state"] == "RUNNING"
